requirements.txt
mkdocs
mkdocs-material
mkdocstrings[python]

pillow
opencv-python
numpy
torch
matplotlib
